vr_dash/requirements.txt
alembic==1.7.4
appnope==0.1.2
autopep8==1.6.0
backcall==0.2.0
backpack==0.1
backports.entry-points-selectable==1.1.0
blinker==1.4
certifi==2021.10.8
cleo==0.6.1
click==8.0.3
debugpy==1.4.3
decorator==5.1.0
distlib==0.3.3
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
Faker==0.8.18
filelock==3.3.1
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
flask-orator==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
idna==3.3
inflection==0.3.1
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jupyter-client==7.0.2
jupyter-core==4.7.1
lazy-object-proxy==1.6.0
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
orator==0.9.9
parso==0.8.2
pastel==0.2.1
pendulum==1.5.1
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2021.5.29
platformdirs==2.4.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyaml==16.12.2
pycodestyle==2.8.0
Pygments==2.10.0
pylev==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.2.1
simplejson==3.17.5
six==1.16.0
SQLAlchemy==1.4.26
text-unidecode==1.2
toml==0.10.2
tornado==6.1
traitlets==5.1.0
tzlocal==1.5.1
virtualenv==20.8.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.2
WTForms==2.3.3
